+++ src/roo_io/stream/memory_input_stream.h
@@ -12,7 +12,7 @@ class MemoryInputStream : public MultipassInputStream {
 
   bool isOpen() const override { return current_ != nullptr; }
 
-  int read(byte* buf, unsigned int count) override {
+  int read(byte* buf, size_t count) override {
     if (current_ == nullptr) return -1;
     if (count > end_ - current_) {
       count = end_ - current_;
